# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "c-plane API",
    "description": "Control plane API for managing projects, environments, containers, and databases",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/s3-access-tokens/resolve/{access_key}": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "resolve_access_token",
        "parameters": [
          {
            "name": "access_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedS3AccessToken"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ]
      }
    },
    "/internal/s3-providers/{provider_id}/credentials": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "provider_credentials",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ProviderCredentials"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "details": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ResolvedS3AccessToken": {
        "type": "object",
        "required": [
          "credential_id",
          "prefix",
          "bucket_permissions"
        ],
        "properties": {
          "bucket_permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedS3BucketPermission"
            }
          },
          "credential_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "prefix": {
            "type": "string"
          }
        }
      },
      "ResolvedS3BucketPermission": {
        "type": "object",
        "required": [
          "bucket_id",
          "bucket_name",
          "physical_bucket_name",
          "region",
          "provider_id",
          "can_read",
          "can_write",
          "is_deleting"
        ],
        "properties": {
          "bucket_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foundation bucket ID."
          },
          "bucket_name": {
            "type": "string",
            "description": "Logical S3 bucket name presented to the credential."
          },
          "can_read": {
            "type": "boolean"
          },
          "can_write": {
            "type": "boolean"
          },
          "is_deleting": {
            "type": "boolean"
          },
          "physical_bucket_name": {
            "type": "string"
          },
          "provider_id": {
            "type": "string",
            "format": "uuid"
          },
          "region": {
            "type": "string"
          }
        }
      },
...
      "S3ProviderCredentials": {
        "type": "object",
        "required": [
          "access_key_id",
          "endpoint_url",
          "name"
        ],
        "properties": {
          "access_key_id": {
            "type": "string"
          },
          "endpoint_url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "provider_region": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key with the operation's required scope."
      },
      "registryBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "serviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cplane-token",
        "description": "Internal service token."
      }
    }
  },
...
}